// src/types/insurance.ts
export type InsurancePolicy = {
  id: string
  source?: string
  sourceLabel?: string
  manufacturer?: string
  company?: string
  mainBranch?: string
  secondaryBranch?: string
  itemClass?: string
  policyNumber?: string
  productType?: string
  insuranceType?: string
  policyName?: string
  planName?: string
  periodText?: string
  premium?: number
  premiumText?: string
  premiumType?: string
  status?: string
  startDate?: string
  endDate?: string
  coverageAmount?: number
  coverageAmountText?: string
  exportDate?: string
  moreDetails?: string
}